+++ CHANGELOG.md
@@ -5,9 +5,23 @@ All notable changes to the Befunge VS Code extension will be documented in this 
 The format of this changelog is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 1.1.0 - 2018-01-21
+
+### Added
+
+* Alignment guides from the directional characters (`>`, `v`, `<`, `^`, `?`)
+
+* Configuration to enable/disable alignment guides, and set the color of them per theme.
+
+* Basic testing of the renderer
+
+### Changed
+
+* Changed name from *Befunge Syntax Highlighting* to *Befunge*, to reflect the additional features.
+
 ## 1.0.0 (Initial Release) - 2018-01-20
 
 ### Added
 
-- Befunge-93 syntax awareness
-- Befunge-98 syntax awareness
+* Befunge-93 syntax awareness
+* Befunge-98 syntax awareness